requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
colorama>=0.4.6
aiohttp>=3.9.0
dnspython>=2.4.0
tqdm>=4.66.0
PyYAML>=6.0.1
urllib3>=2.0.0
lxml>=4.9.0
